Move FilterProxyModel and RealDelegate into a utility module
